+++ database/migrations/2021_10_13_131327_create_profits_table.php
@@ -15,7 +15,7 @@ class CreateProfitsTable extends Migration
     {
         Schema::create('profits', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->float('profit_rate');
+            $table->decimal('profit_rate', 10, 10);
             $table->timestamps();
         });
     }
